feat: FastAPI 검색, 매칭, 유사도 뼈대 수정 1. 기존 라우터 3개에 통합, 구조 유지 2. 필요없는 엔드포인트 삭제 3. schemas 데이터 일관성 불일치 수정 4. 중복 import 삭제
