+++ apps/pre-processing-service/app/api/endpoints/keywords.py
@@ -2,11 +2,8 @@
 from ...service.keyword_service import keyword_search
 
 from fastapi import APIRouter
-from app.decorators.logging import log_api_call
-from ...errors.CustomException import *
-from fastapi import APIRouter
 from ...errors.CustomException import  *
-from ...model.schemas import RequestNaverSearch, ResponseNaverSearch, RequestSadaguValidate, ResponsetSadaguValidate
+from ...model.schemas import RequestNaverSearch, ResponseNaverSearch
 
 # 이 파일만의 독립적인 라우터를 생성합니다.
 router = APIRouter()
@@ -32,7 +29,3 @@ async def search(request: RequestNaverSearch):
     """
     response_data= await keyword_search(request)
     return response_data
-
-@router.post("/ssadagu/validate",response_model=ResponseNaverSearch)
-async def ssadagu_validate(request: RequestNaverSearch):
-    return ResponseNaverSearch()
