--- apps/pre-processing-service/app/api/endpoints/match.py
@@ -1,9 +1,0 @@
-from fastapi import APIRouter
-from app.service.match_service import match_products
-from app.model.schemas import RequestSadaguMatch, ResponseSadaguMatch
-
-router = APIRouter()
-
-@router.post("/", response_model=ResponseSadaguMatch)
-async def match(request: RequestSadaguMatch):
-    return match_products(request)
